# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atlas - Project-based technical learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/projects:
    get:
      operationId: listUserProjects
      tags: [user-projects]
      summary: List authenticated user's enrolled projects
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User's project enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProjectSummary"
  /user/projects/{projectId}/enroll:
    post:
      operationId: enrollProject
      tags: [user-projects]
      summary: Enroll in a project
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrollment created or existing enrollment returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProject"
        "403":
          description: Project requires Pro subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/projects/{projectId}/progress:
    get:
      operationId: getUserProjectProgress
      tags: [user-projects]
      summary: Get detailed progress for a user project
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User project progress with step completions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProjectProgress"
  /user/projects/{projectId}/steps/{stepId}/submit:
    post:
      operationId: submitStep
      tags: [steps]
      summary: Submit a step for grading
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitStepBody"
      responses:
        "200":
          description: Grading result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingResult"
        "404":
          description: Step not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/projects/{projectId}/steps/{stepId}/hint:
    post:
      operationId: requestStepHint
      tags: [steps]
      summary: Request a hint for a step
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HintRequestBody"
      responses:
        "200":
          description: AI-generated hint (streaming SSE)
          content:
            text/event-stream:
              schema:
                type: string
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get authenticated user's profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user learning statistics and XP
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        tier:
          type: string
          enum: [free, pro]
        xpReward:
          type: integer
        estimatedHours:
          type: number
        stepCount:
          type: integer
        enrolledCount:
          type: integer
        completionRate:
          type: number
        tags:
          type: array
          items:
            type: string
        position:
          type: integer
      required:
        - id
        - slug
        - title
        - description
        - difficulty
        - tier
        - xpReward
        - estimatedHours
        - stepCount
        - enrolledCount
        - completionRate
        - tags
        - position
...
    ProjectStep:
      type: object
      properties:
        id:
          type: string
        position:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [code_python, code_sql, multiple_choice, true_false, short_answer, file_upload, diagram, concept_check, reflection]
        starterCode:
          type: string
        expectedOutput:
          type: string
        hints:
          type: array
          items:
            type: string
        xpReward:
          type: integer
        isLocked:
          type: boolean
      required:
        - id
        - position
        - title
        - description
        - type
        - xpReward
        - isLocked
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            longDescription:
              type: string
            learningObjectives:
              type: array
              items:
                type: string
            prerequisites:
              type: array
              items:
                type: string
            steps:
              type: array
              items:
                $ref: "#/components/schemas/ProjectStep"
            domainSlug:
              type: string
            domainName:
              type: string
    UserProject:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed, abandoned]
        currentStepPosition:
          type: integer
        earnedXp:
          type: integer
        completedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - userId
        - status
        - currentStepPosition
        - earnedXp
    UserProjectSummary:
      allOf:
        - $ref: "#/components/schemas/UserProject"
        - type: object
          properties:
            project:
              $ref: "#/components/schemas/ProjectSummary"
    UserProjectProgress:
      allOf:
        - $ref: "#/components/schemas/UserProject"
        - type: object
          properties:
            project:
              $ref: "#/components/schemas/ProjectDetail"
            stepCompletions:
              type: array
              items:
                $ref: "#/components/schemas/StepCompletion"
            progressPercent:
              type: number
    StepCompletion:
      type: object
      properties:
        id:
          type: string
        stepId:
          type: string
        userProjectId:
          type: string
        status:
          type: string
          enum: [pending, passed, failed, skipped]
        attempt:
          type: integer
        submission:
          type: string
        feedback:
          type: string
        xpEarned:
          type: integer
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - stepId
        - userProjectId
        - status
        - attempt
    SubmitStepBody:
      type: object
      properties:
        submission:
          type: string
          description: Code, answer, or other submission content
        submissionType:
          type: string
          enum: [code, text, choice, file]
      required:
        - submission
        - submissionType
    GradingResult:
      type: object
      properties:
        status:
          type: string
          enum: [passed, failed]
        feedback:
          type: string
        xpEarned:
          type: integer
        attempt:
          type: integer
        isFirstPass:
          type: boolean
        nextStepId:
          type: string
        projectComplete:
          type: boolean
        stdout:
          type: string
        stderr:
          type: string
        executionTimeMs:
          type: integer
      required:
        - status
        - feedback
        - xpEarned
        - attempt
    HintRequestBody:
      type: object
      properties:
        currentCode:
          type: string
        errorMessage:
          type: string
        hintIndex:
          type: integer
      required:
        - hintIndex
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        tier:
          type: string
          enum: [free, pro]
        totalXp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - tier
        - totalXp
        - level
        - streak
    UpdateUserProfileBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
    UserStats:
      type: object
      properties:
        totalXp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        projectsCompleted:
          type: integer
        projectsInProgress:
          type: integer
        lessonsCompleted:
          type: integer
        stepsCompleted:
          type: integer
        rank:
          type: integer
        xpToNextLevel:
          type: integer
        weeklyXp:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              xp:
                type: integer
      required:
        - totalXp
        - level
        - streak
        - projectsCompleted
        - projectsInProgress
        - lessonsCompleted
        - stepsCompleted
...