# lib/api-zod/src/generated/types/submitStepBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Project-based technical learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitStepBodySubmissionType } from "./submitStepBodySubmissionType";

export interface SubmitStepBody {
  /** Code, answer, or other submission content */
  submission: string;
  submissionType: SubmitStepBodySubmissionType;
}